backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
python-dotenv==1.1.0
Werkzeug==3.1.3

# Added for database support
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.29